Refactor code for tile upgrades and Seed item
Updated the code to fix the tile upgrade functionality. Changed the way coordinates are accessed in the upgradeTile action in upgrade.blade.php. Also, refactored WoodTile class in WoodTile.php, to improve modularity with shared traits.

In Seed.php, added check for tile's type before useOn method and set elevation based on the current tile's elevation instead of a static value. This change is made for accuracy and prevents the Seed item from being used on unsuitable tiles.