"Refactor code by removing unused imports and update classes"
This commit addresses the need for cleanup in many PHP files of the application. Numerous unused imports across different components were removed to increase readability and maintainability. The BaseLayer in MapCanvasController class is also updated, merging the ElevationLayer, IslandLayer, and BiomeLayer into a single BiomeLayer being add to the BaseLayer. Furthermore, the getElevation method is in DebugTile class to provide additional data about the tile's elevation. These changes aim to enhance the application's overall performance and simplicity.