Update map canvas styling and add zoom feature
Updated the styling of the map canvas view, by introducing css classes instead of inline styles. Also increased the dimensions and reduced the pixel size in the MapCanvasController for finer map details.
Added a zoom feature to the map. Map can now be zoomed in/out using the mouse wheel while holding the shift key. Refactored map initialization code into a draw function to enable re-rendering the map on zoom.